PF-153 Add WM chart values for spend and google folder for creating projects (#217)
This will enable us to configure WM with a spend profile id/billing account id and a google folder to create projects in.

While I'm not crazy about the continued proliferation of helm chart values & environment variables, we are still awaiting a long term solution for passing the application.yaml through the helm chart. So good enough for now.